docs(DST-1384): simplify the bulk-actions example
Code-only cleanup, no behavior change:

- Merge the three URL-param hooks (usePagination/useSearch/useStatusFilter)
  into one useEventsParams.ts, so the page-bounded "clear selection on
  scope change" rule is documented and read in one place
- Replace the hand-rolled <div role="separator"> in the ActionBar with the
  Marigold <Divider orientation="vertical"> so the example composes only
  Marigold components
- Factor the repeated success/partial-failure toast + selection handling
  out of the four mutations into one report() helper, and derive the
  failure reason from the server result instead of hardcoding it
- Flatten the pagination layout to a Marigold <Inline>, dropping the
  three-column grid and its empty spacer div
